+++ packages/@webex/plugin-meetings/src/meeting/index.ts
@@ -190,6 +190,7 @@ export const MEDIA_UPDATE_TYPE = {
  * @property {String} [meetingQuality.remote]
  * @property {Boolean} [rejoin]
  * @property {Boolean} [enableMultistream]
+ * @property {String} [correlationId]
  */
 
 /**
@@ -1205,9 +1206,11 @@ export default class Meeting extends StatelessWebexPlugin {
   public async fetchMeetingInfo({
     password = null,
     captchaCode = null,
+    extraParams = {},
   }: {
     password?: string;
     captchaCode?: string;
+    extraParams?: Record<string, any>;
   }) {
     // when fetch meeting info is called directly by the client, we want to clear out the random timer for sdk to do it
     if (this.fetchMeetingInfoTimeoutId) {
@@ -1241,7 +1244,8 @@ export default class Meeting extends StatelessWebexPlugin {
         captchaInfo,
         // @ts-ignore - config coming from registerPlugin
         this.config.installedOrgID,
-        this.locusId
+        this.locusId,
+        extraParams
       );
 
       this.parseMeetingInfo(info, this.destination);
@@ -1394,6 +1398,20 @@ export default class Meeting extends StatelessWebexPlugin {
           throw error;
         })
     );
+  }
+
+  /**
+   * Posts metrics event for this meeting. Allows the app to send Call Analyzer events.
+   * @param {String} eventName - Call Analyzer event, see eventType in src/metrics/config.ts for possible values
+   * @public
+   * @memberof Meeting
+   * @returns {Promise}
+   */
+  public postMetrics(eventName: string) {
+    Metrics.postEvent({
+      event: eventName,
+      meeting: this,
+    });
   }
 
   /**
@@ -4285,9 +4303,16 @@ export default class Meeting extends StatelessWebexPlugin {
       joinSuccess = resolve;
     });
 
+    if (options.correlationId) {
+      this.setCorrelationId(options.correlationId);
+      LoggerProxy.logger.log(
+        `Meeting:index#join --> Using a new correlation id from app ${this.correlationId}`
+      );
+    }
+
     if (!this.hasJoinedOnce) {
       this.hasJoinedOnce = true;
-    } else {
+    } else if (!options.correlationId) {
       LoggerProxy.logger.log(
         `Meeting:index#join --> Generating a new correlation id for meeting ${this.id}`
       );