+++ packages/@webex/plugin-meetings/test/unit/spec/meeting/index.js
@@ -3010,6 +3010,7 @@ describe('plugin-meetings', () => {
         const FAKE_CAPTCHA_AUDIO_URL = 'http://captchaaudio';
         const FAKE_CAPTCHA_REFRESH_URL = 'http://captcharefresh';
         const FAKE_INSTALLED_ORG_ID = '123456';
+        const FAKE_EXTRA_PARAMS = {mtid: 'm9fe0afd8c435e892afcce9ea25b97046', joinTXId: 'TSmrX61wNF'};
         const FAKE_MEETING_INFO = {
           conversationUrl: 'some_convo_url',
           locusUrl: 'some_locus_url',
@@ -3043,6 +3044,7 @@ describe('plugin-meetings', () => {
           await meeting.fetchMeetingInfo({
             password: FAKE_PASSWORD,
             captchaCode: FAKE_CAPTCHA_CODE,
+            extraParams: FAKE_EXTRA_PARAMS,
           });
 
           assert.calledWith(
@@ -3051,7 +3053,9 @@ describe('plugin-meetings', () => {
             FAKE_TYPE,
             FAKE_PASSWORD,
             {code: FAKE_CAPTCHA_CODE, id: FAKE_CAPTCHA_ID},
-            FAKE_INSTALLED_ORG_ID
+            FAKE_INSTALLED_ORG_ID,
+            meeting.locusId,
+            FAKE_EXTRA_PARAMS
           );
 
           assert.calledWith(meeting.parseMeetingInfo, {body: FAKE_MEETING_INFO}, FAKE_DESTINATION);
@@ -3093,7 +3097,10 @@ describe('plugin-meetings', () => {
             FAKE_DESTINATION,
             FAKE_TYPE,
             null,
-            null
+            null,
+            undefined,
+            meeting.locusId,
+            {}
           );
 
           // parseMeeting info
@@ -3487,6 +3494,17 @@ describe('plugin-meetings', () => {
             {file: 'meeting/index', function: 'mediaNegotiatedEvent'},
             'media:negotiated'
           );
+        });
+      });
+
+      describe('#postMetrics', () => {
+        it('should have #postMetrics', () => {
+          assert.exists(meeting.postMetrics);
+        });
+
+        it('should trigger `postMetrics`', async () => {
+          await meeting.postMetrics(eventType.LEAVE);
+          assert.calledWithMatch(Metrics.postEvent, {event: eventType.LEAVE});
         });
       });